refactor: renamed an overly long folder name
